+++ blocklets/ai-studio/src/components/component-select/ComponentSelect.tsx
@@ -1,26 +1,10 @@
-import { useCurrentProject } from '@app/contexts/project';
 import { getComponents } from '@app/libs/components';
 import { REMOTE_REACT_COMPONENT } from '@app/libs/constants';
-import { useProjectStore } from '@app/pages/project/yjs-state';
+import { getOptimizedImageAbsUrl } from '@app/libs/media';
+import Empty from '@arcblock/ux/lib/Empty';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
-import { parseIdentity } from '@blocklet/ai-runtime/common/aid';
-import {
-  OutputVariable,
-  OutputVariableYjs,
-  RuntimeOutputVariable,
-  fileFromYjs,
-  isAssistant,
-  outputVariableFromYjs,
-  outputVariablesToJsonSchema,
-  variableFromYjs,
-} from '@blocklet/ai-runtime/types';
-import {
-  CurrentMessageOutputProvider,
-  CurrentMessageProvider,
-  RuntimeDebug,
-  RuntimeProvider,
-} from '@blocklet/aigne-sdk/components/ai-runtime';
-import { CustomComponentRenderer, CustomComponentRendererProvider } from '@blocklet/pages-kit/components';
+import { OutputVariableYjs, RuntimeOutputVariable } from '@blocklet/ai-runtime/types';
+import { RuntimeDebug } from '@blocklet/aigne-sdk/components/ai-runtime';
 import {
   Alert,
   Box,
@@ -40,15 +24,45 @@ import {
   styled,
   useRadioGroup,
 } from '@mui/material';
+import Ajv from 'ajv';
 import stringify from 'json-stable-stringify';
 import { pick } from 'lodash';
-import { nanoid } from 'nanoid';
-import { ComponentProps, ReactNode, Suspense, useEffect, useMemo, useState } from 'react';
+import { ComponentProps, Suspense, useEffect, useMemo, useState } from 'react';
 import { useAsync } from 'react-use';
 
 import ErrorBoundary from '../error/error-boundary';
 import { generateFakeProps } from './fake-props';
 import { useAIGNEApiProps } from './get-agent';
+
+const ajv = new Ajv({ strict: false });
+
+function outputToJsonSchema(output: Record<string, any>) {
+  if (!output || output.from?.type === 'input') return {};
+  function convert(data: Record<string, any>) {
+    const ans: Record<string, any> = { type: data.type ?? 'string' };
+
+    if (data.type === 'object' && data.properties) {
+      ans.properties = {};
+      for (const [, v] of Object.entries(data.properties)) {
+        const value = v as { data: Record<string, any> };
+        ans.properties[value.data.name] = convert(value.data);
+        if (value.data.required) {
+          ans.required = ans.required ?? [];
+          ans.required.push(value.data.name);
+        }
+      }
+      return ans;
+    }
+
+    if (data.type === 'array' && data.element) {
+      ans.items = convert(data.element);
+      return ans;
+    }
+
+    return ans;
+  }
+  return convert(output);
+}
 
 export interface ComponentSelectValue {
   blockletDid?: string;
@@ -91,6 +105,9 @@ export default function ComponentSelect({
         blockletDid: x.blocklet?.did,
         id: x.id,
         name: x.name,
+        description: x.description,
+        previewImage: x.previewImage,
+        aigneOutputValueSchema: x.aigneOutputValueSchema,
         group: t('buildIn'),
       })),
       ...(openComponents ?? []).map((x) => ({
@@ -104,9 +121,26 @@ export default function ComponentSelect({
         group: t('remote'),
       })),
     ];
-  }, [customComponents, openComponents]);
+  }, [customComponents, openComponents, t]);
 
   const componentsMap = useMemo(() => Object.fromEntries(components.map((i) => [i.value, i])), [components]);
+
+  const validatedComponentIds = useMemo(() => {
+    if (!output.type) return new Set(components.map((i) => i.id));
+
+    const outputSchema = outputToJsonSchema(output);
+    const outputSchemaFakedData = generateFakeProps(outputSchema);
+    const validatedComponents =
+      components?.filter((i) => {
+        const componentSchema = (i as any).aigneOutputValueSchema;
+        if (!componentSchema) return true;
+        const validate = ajv.compile(componentSchema);
+        return validate({ outputValue: outputSchemaFakedData });
+      }) ?? [];
+    return new Set(validatedComponents.map((i) => i.id).filter(Boolean));
+  }, [components, output]);
+
+  const validatedComponents = components.filter((i) => validatedComponentIds.has(i.id));
 
   return (
     <Grid
@@ -119,7 +153,7 @@ export default function ComponentSelect({
       }}
       container
       spacing={2}>
-      {components.map((i) => (
+      {validatedComponents.map((i) => (
         <Grid item key={i.value} xs={12} sm={6} md={4}>
           <ComponentSelectItem output={output} aid={aid} customComponent={i} />
         </Grid>
@@ -141,7 +175,14 @@ function ComponentSelectItem({
 }: {
   output: OutputVariableYjs;
   aid: string;
-  customComponent: { value: string; blockletDid?: string; id: string; name?: string; componentProperties?: any };
+  customComponent: {
+    value: string;
+    blockletDid?: string;
+    id: string;
+    name?: string;
+    description?: string;
+    componentProperties?: any;
+  };
 }) {
   const radioGroup = useRadioGroup();
   const checked = radioGroup?.value === customComponent.value;
@@ -180,6 +221,13 @@ function ComponentSelectItem({
         <Typography sx={{ m: 1 }} noWrap>
           {customComponent.name || t('unnamed')}
         </Typography>
+        <Typography
+          sx={{ m: 1, WebkitLineClamp: 2, overflowWrap: 'break-word' }}
+          className="multi-line-ellipsis"
+          variant="caption"
+          title={customComponent.description}>
+          {customComponent.description}
+        </Typography>
       </Box>
     </Card>
   );
@@ -198,123 +246,38 @@ function ItemPreviewer({
     return <PageComponentPreviewer aid={aid} customComponent={customComponent} />;
   }
 
-  return <ComponentPreviewer output={output} aid={aid} customComponent={customComponent} />;
+  return <ComponentPreviewImagePreviewer customComponent={customComponent} />;
 }
 
-function ComponentPreviewer({
-  output,
-  aid,
+function ComponentPreviewImagePreviewer({
   customComponent,
 }: {
-  output: OutputVariableYjs;
-  aid: string;
-  customComponent: { blockletDid?: string; id: string; name?: string; componentProperties?: any };
+  customComponent: {
+    id: string;
+    name?: string;
+    description?: string;
+    previewImage?: string;
+  };
 }) {
-  const { projectId, projectRef } = useCurrentProject();
-  const { agentId } = useMemo(() => parseIdentity(aid, { rejectWhenError: true }), [aid]);
-  const { getFileById, getVariables } = useProjectStore(projectId, projectRef);
-
-  const agent = getFileById(agentId);
-  if (!agent) throw new Error(`No such agent ${agentId}`);
-  if (!isAssistant(agent)) throw new Error(`Invalid agent file type ${agentId}`);
-
-  const fakeMessage = useMemo<ComponentProps<typeof CurrentMessageProvider>['message']>(() => {
-    const { agentId } = parseIdentity(aid, { rejectWhenError: true });
-
-    const file = fileFromYjs(agent);
-    if (!isAssistant(file)) throw new Error(`Invalid agent file type ${agentId}`);
-
-    const outputSchema = outputVariablesToJsonSchema(file, {
-      variables: getVariables().variables?.map(variableFromYjs) ?? [],
-      includeRuntimeOutputVariables: true,
-      includeFaker: true,
-    });
-
-    const output = outputSchema ? generateFakeProps(outputSchema) : {};
-
-    return {
-      id: nanoid(),
-      aid,
-      agentId,
-      sessionId: nanoid(),
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      outputs: {
-        objects: [output as any],
-      },
-    };
-  }, [aid]);
-
-  const api = useAIGNEApiProps({
-    apiUniqueKey: `ComponentSelectPreviewer-${customComponent.id}-${output.name}`,
-    customComponent,
-  });
-
-  const customComponentCtx = useMemo(
-    () => ({
-      customRendererComponent: FakeMessageOutputValueProvider,
-      customRendererComponentProps: { output: outputVariableFromYjs(output) },
-    }),
-    []
-  );
-
-  console.warn(customComponentCtx);
-
+  const { t } = useLocaleContext();
+  const { previewImage, name } = customComponent;
   return (
     <Box position="relative" paddingBottom="100%">
       <Box position="absolute" left={0} top={0} right={0} bottom={0}>
-        <ErrorBoundary FallbackComponent={ErrorView}>
-          <Suspense
-            fallback={
-              <Box textAlign="center" my={2}>
-                <CircularProgress size={24} />
-              </Box>
-            }>
-            <PreviewerContent>
-              <Box
-                sx={{
-                  position: 'absolute',
-                  width: '100%',
-                  top: '50%',
-                  left: '50%',
-                  transform: 'translate(-50%,-50%)',
-                }}>
-                <RuntimeProvider aid={aid} working ApiProps={api}>
-                  <CurrentMessageProvider message={fakeMessage}>
-                    <CustomComponentRendererProvider value={customComponentCtx}>
-                      <CustomComponentRenderer componentId={customComponent.id} />
-                    </CustomComponentRendererProvider>
-                  </CurrentMessageProvider>
-                </RuntimeProvider>
-              </Box>
-            </PreviewerContent>
-          </Suspense>
-        </ErrorBoundary>
+        {previewImage ? (
+          <Box
+            component="img"
+            sx={{ width: '100%', height: '100%', objectFit: 'contain' }}
+            src={getOptimizedImageAbsUrl(previewImage)}
+            alt={name}
+          />
+        ) : (
+          <Empty size={30}>
+            <Box sx={{ fontSize: '16px' }}>{t('noPreviewImage')}</Box>
+          </Empty>
+        )}
       </Box>
     </Box>
-  );
-}
-
-function FakeMessageOutputValueProvider({
-  output,
-  children,
-  Component,
-}: {
-  output: OutputVariable;
-  children?: ReactNode;
-  // @ts-ignore
-  Component: CustomComponentType;
-}) {
-  const fakeValue = useMemo(
-    () =>
-      Component.outputValueSchema ? (generateFakeProps(Component.outputValueSchema) as any)?.outputValue : undefined,
-    [Component.outputValueSchema]
-  );
-
-  return (
-    <CurrentMessageOutputProvider output={output} outputValue={fakeValue}>
-      {children}
-    </CurrentMessageOutputProvider>
   );
 }