+++ blocklets/ai-studio/src/locales/en.ts
@@ -878,6 +878,7 @@ export default flatten({
   multiple: 'Multiple',
   dropdown: 'Dropdown',
   checkbox: 'Checkbox',
+  noPreviewImage: 'No Preview Image',
   verifyVC: 'Verify VC',
   vcItem: 'VC Name',
   trustedIssuers: 'Trusted Issuers',